Add virtio-blk DMA, unlock block I/O, and pin the identity memory map.
Per-device locks and 64-bit LBAs stop ATA PIO from freezing the system.
QEMU disks use virtio-blk (MSI-X + DMA). Reserved physical ranges live in
memlayout.h so kernel BSS cannot clobber the frame stack again.

+++ ics-os/kernel/devmgr/dex32_devmgr.h
@@ -129,7 +129,7 @@ devmgr_generic hdr;
 DWORD (*init)();
 int   (*complete)(DWORD handle);
 void  (*close)(DWORD handle);
-DWORD (*request)(int,int,DWORD,DWORD,void*);
+DWORD (*request)(int,int,u64,DWORD,void*);
 } devmgr_iomgr;
 
 typedef struct _devmgr_mem {
@@ -175,8 +175,8 @@ devmgr_generic hdr;
 int (*init_device)();
 int (*invalidate_cache)();
 int (*getcache)(char *buf,DWORD sectornumber,DWORD numblocks);
-int (*read_block)(int block,char *blockbuff, DWORD numblocks);
-int (*write_block)(int block,char *blockbuff,DWORD numblocks);
+int (*read_block)(u64 block,char *blockbuff, DWORD numblocks);
+int (*write_block)(u64 block,char *blockbuff,DWORD numblocks);
 int (*total_blocks) ();
 int (*flush_device)();
 int (*get_block_size)();
